// admin_frontend/src/lib/api/models/ProductUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductUpdate = {
    title?: (string | null);
    slug?: (string | null);
    tea_type?: (string | null);
    description?: (string | null);
    lore_description?: (string | null);
    brewing_guide?: (Record<string, any> | null);
    category_id?: (number | null);
    is_active?: (boolean | null);
    seo_title?: (string | null);
    seo_description?: (string | null);
    seo_keywords?: (string | null);
};